+++ src/app/views/Alerts/actions/detail.spec.ts
@@ -17,7 +17,6 @@
  */
 
 // Vendor
-import * as React from 'react';
 import * as sinon from 'sinon';
 import * as chai from 'chai';